requirements.txt
webcolors~=1.13
pygame~=2.5.0
future~=0.18.3
numpy~=1.25.2
Pillow~=10.0.0